add second example of custom UI
